Move supplier form parts to specific file and make them be added through provides
